+++ retrocache/src/test/java/com/andiag/retrocache/SimpleTest.java
@@ -1,6 +1,5 @@
 package com.andiag.retrocache;
 
-import com.andiag.retrocache.annotations.Caching;
 import com.andiag.retrocache.interfaces.CachedCall;
 import com.andiag.retrocache.utils.MainThreadExecutor;
 import com.andiag.retrocache.utils.MockCachingSystem;
@@ -21,9 +20,7 @@ import retrofit2.CallAdapter;
 import retrofit2.Callback;
 import retrofit2.Response;
 import retrofit2.Retrofit;
-import retrofit2.http.Body;
 import retrofit2.http.GET;
-import retrofit2.http.POST;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -69,7 +66,8 @@ public class SimpleTest {
             @Override
             public void onResponse(Call<String> call, Response<String> response) {
                 assertEquals(response.body(), "VERY_BASIC_BODY");
-                assertNotNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+                assertNotNull(mMockCachingSystem.get(
+                        Utils.urlToKey(call.request().method(), call.request().url())));
             }
 
             @Override
@@ -93,7 +91,8 @@ public class SimpleTest {
             @Override
             public void onResponse(Call<String> call, Response<String> response) {
                 assertEquals(response.body(), "VERY_BASIC_BODY");
-                assertNotNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+                assertNotNull(mMockCachingSystem.get(
+                        Utils.urlToKey(call.request().method(), call.request().url())));
                 latch.countDown();
             }
 
@@ -104,7 +103,8 @@ public class SimpleTest {
         });
         latch.await(1, TimeUnit.SECONDS);
         call.remove();
-        assertNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+        assertNull(mMockCachingSystem.get(Utils.urlToKey(
+                call.request().method(), call.request().url())));
     }
 
     @Test
@@ -122,7 +122,8 @@ public class SimpleTest {
             @Override
             public void onResponse(Call<String> call, Response<String> response) {
                 assertEquals(response.body(), "VERY_BASIC_BODY");
-                assertNotNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+                assertNotNull(mMockCachingSystem.get(
+                        Utils.urlToKey(call.request().method(), call.request().url())));
                 latch.countDown();
             }
 
@@ -140,53 +141,8 @@ public class SimpleTest {
             @Override
             public void onResponse(Call<String> call, Response<String> response) {
                 assertEquals(response.body(), "VERY_BASIC_REFRESHED_BODY");
-                assertNotNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
-            }
-
-            @Override
-            public void onFailure(Call<String> call, Throwable t) {
-                fail("Failure executing the request: " + t.getMessage());
-            }
-        });
-    }
-
-    @Test
-    public void disabledCache() {
-        /* Set up the mock webserver */
-        MockResponse resp = new MockResponse().setBody("VERY_BASIC_BODY");
-        mServer.enqueue(resp);
-
-        DemoService demoService = mRetrofit.create(DemoService.class);
-        CachedCall<String> call = demoService.getHomeDisablingCache();
-
-        call.enqueue(new Callback<String>() {
-            @Override
-            public void onResponse(Call<String> call, Response<String> response) {
-                assertEquals(response.body(), "VERY_BASIC_BODY");
-                assertNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
-            }
-
-            @Override
-            public void onFailure(Call<String> call, Throwable t) {
-                fail("Failure executing the request: " + t.getMessage());
-            }
-        });
-    }
-
-    @Test
-    public void cachedPost() {
-        /* Set up the mock webserver */
-        MockResponse resp = new MockResponse().setBody("VERY_BASIC_BODY");
-        mServer.enqueue(resp);
-
-        DemoService demoService = mRetrofit.create(DemoService.class);
-        CachedCall<String> call = demoService.postHome("SIMPLE_BODY");
-
-        call.enqueue(new Callback<String>() {
-            @Override
-            public void onResponse(Call<String> call, Response<String> response) {
-                assertEquals(response.body(), "VERY_BASIC_BODY");
-                assertNotNull(mMockCachingSystem.get(Utils.urlToKey(call.request().url())));
+                assertNotNull(mMockCachingSystem.get(
+                        Utils.urlToKey(call.request().method(), call.request().url())));
             }
 
             @Override
@@ -200,14 +156,6 @@ public class SimpleTest {
 
         @GET("/")
         CachedCall<String> getHome();
-
-        @Caching(enabled = false)
-        @GET("/")
-        CachedCall<String> getHomeDisablingCache();
-
-        @Caching
-        @POST("/")
-        CachedCall<String> postHome(@Body String body);
 
     }
 }